adjust package
