+++ src/main/java/br/com/fiap/hacka/notificacao/domain/model/IEnumLabel.java
@@ -1,7 +1,7 @@
-package br.com.fiap.hacka.domain.model;
+package br.com.fiap.hacka.notificacao.domain.model;
 
 
-import br.com.fiap.hacka.infrastructure.utils.MensagensUtils;
+import br.com.fiap.hacka.notificacao.infrastructure.utils.MensagensUtils;
 
 public interface IEnumLabel<E extends Enum<E>>{
     default String getDescricao(){
