+++ src/main/java/br/com/fiap/hacka/notificacao/input/controller/NotificacaoApi.java
@@ -1,11 +1,9 @@
-package br.com.fiap.hacka.adapter.input.controller;
+package br.com.fiap.hacka.notificacao.input.controller;
 
-import br.com.fiap.hacka.adapter.input.dto.NotificacaoDto;
+import br.com.fiap.hacka.notificacao.input.dto.NotificacaoDto;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestBody;
-
-import java.util.List;
 
 public interface NotificacaoApi {
     ResponseEntity<Void> sendWebhookMessage(@Valid @RequestBody NotificacaoDto notificacaoDto);
